+++ lib/screens/add_meal_plan.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database_old.dart';
 import 'package:recipes/models/meal_plan.dart';
 import 'package:recipes/models/recipe.dart';
 import 'package:recipes/screens/meal_plan.dart';
