+++ lib/screens/meal_plan.dart
@@ -4,7 +4,7 @@ import 'package:get_it/get_it.dart';
 import 'package:recipes/blocs/meal_plan/bloc.dart';
 import 'package:recipes/blocs/meal_plan/events.dart';
 import 'package:recipes/blocs/meal_plan/state.dart';
-import 'package:recipes/database.dart';
+import 'package:recipes/database_old.dart';
 import 'package:recipes/helpers/number_formatters.dart';
 import 'package:recipes/models/grocery.dart';
 import 'package:recipes/helpers/add_ingredients_to_groceries.dart';
